// src/app/grpc/model/accountBalance_pb.d.ts
export class AccountBalance extends jspb.Message {
  getAccountaddress(): string;
  setAccountaddress(value: string): void;

  getBlockheight(): number;
  setBlockheight(value: number): void;

  getSpendablebalance(): string;
  setSpendablebalance(value: string): void;

  getBalance(): string;
  setBalance(value: string): void;

  getPoprevenue(): string;
  setPoprevenue(value: string): void;

  getLatest(): boolean;
  setLatest(value: boolean): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): AccountBalance.AsObject;
  static toObject(includeInstance: boolean, msg: AccountBalance): AccountBalance.AsObject;
  static serializeBinaryToWriter(message: AccountBalance, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): AccountBalance;
  static deserializeBinaryFromReader(message: AccountBalance, reader: jspb.BinaryReader): AccountBalance;
}

export namespace AccountBalance {
  export type AsObject = {
    accountaddress: string,
    blockheight: number,
    spendablebalance: string,
    balance: string,
    poprevenue: string,
    latest: boolean,
  }
}
// ...
export class GetAccountBalanceResponse extends jspb.Message {
  getAccountbalance(): AccountBalance | undefined;
  setAccountbalance(value?: AccountBalance): void;
  hasAccountbalance(): boolean;
  clearAccountbalance(): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): GetAccountBalanceResponse.AsObject;
  static toObject(includeInstance: boolean, msg: GetAccountBalanceResponse): GetAccountBalanceResponse.AsObject;
  static serializeBinaryToWriter(message: GetAccountBalanceResponse, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): GetAccountBalanceResponse;
  static deserializeBinaryFromReader(message: GetAccountBalanceResponse, reader: jspb.BinaryReader): GetAccountBalanceResponse;
}

export namespace GetAccountBalanceResponse {
  export type AsObject = {
    accountbalance?: AccountBalance.AsObject,
  }
}